[spicedb] Cleanup SpiceDBAuthorizer
