+++ components/server/src/authorization/spicedb-authorizer.ts
@@ -8,11 +8,14 @@ import { v1 } from "@authzed/authzed-node";
 import { log } from "@gitpod/gitpod-protocol/lib/util/logging";
 import { TrustedValue } from "@gitpod/gitpod-protocol/lib/util/scrubbing";
 
-import { inject, injectable } from "inversify";
-import { observeSpicedbClientLatency, spicedbClientLatency } from "../prometheus-metrics";
+import { incSpiceDBRequestsCheckTotal, observeSpicedbClientLatency, spicedbClientLatency } from "../prometheus-metrics";
 import { SpiceDBClientProvider } from "./spicedb";
 import * as grpc from "@grpc/grpc-js";
 import { isFgaChecksEnabled, isFgaWritesEnabled } from "./authorizer";
+import { base64decode } from "@jmondi/oauth2-server";
+import { DecodedZedToken } from "@gitpod/spicedb-impl/lib/impl/v1/impl.pb";
+import { RequestContext } from "node-fetch";
+import { getRequestContext } from "../util/request-context";
 
 async function tryThree<T>(errMessage: string, code: (attempt: number) => Promise<T>): Promise<T> {
     let attempt = 0;
@@ -38,42 +41,43 @@ async function tryThree<T>(errMessage: string, code: (attempt: number) => Promis
     throw new Error("unreachable");
 }
 
-export const SpiceDBAuthorizer = Symbol("SpiceDBAuthorizer");
-export interface SpiceDBAuthorizer {
-    check(
-        req: v1.CheckPermissionRequest,
-        experimentsFields: {
-            userId: string;
-        },
-        forceEnablement?: boolean,
-    ): Promise<CheckResult>;
-    writeRelationships(...updates: v1.RelationshipUpdate[]): Promise<v1.WriteRelationshipsResponse | undefined>;
-    deleteRelationships(req: v1.DeleteRelationshipsRequest): Promise<DeletionResult>;
-    readRelationships(req: v1.ReadRelationshipsRequest): Promise<v1.ReadRelationshipsResponse[]>;
+export function createSpiceDBAuthorizer(clientProvider: SpiceDBClientProvider): SpiceDBAuthorizer {
+    return new SpiceDBAuthorizer(clientProvider, new RequestLocalZedTokenCache());
 }
 
-export interface CheckResult {
+interface CheckResult {
     permitted: boolean;
     checkedAt?: string;
 }
 
-export interface DeletionResult {
+interface DeletionResult {
     relationships: v1.ReadRelationshipsResponse[];
     deletedAt?: string;
 }
 
-@injectable()
-export class SpiceDBAuthorizerImpl implements SpiceDBAuthorizer {
-    constructor(
-        @inject(SpiceDBClientProvider)
-        private readonly clientProvider: SpiceDBClientProvider,
-    ) {}
+export class SpiceDBAuthorizer {
+    constructor(private readonly clientProvider: SpiceDBClientProvider, private readonly tokenCache: ZedTokenCache) {}
 
     private get client(): v1.ZedPromiseClientInterface {
         return this.clientProvider.getClient();
     }
 
-    async check(
+    public async check(
+        req: v1.CheckPermissionRequest,
+        experimentsFields: { userId: string },
+        forceEnablement?: boolean,
+    ): Promise<boolean> {
+        req.consistency = await this.tokenCache.consistency(req.resource);
+        incSpiceDBRequestsCheckTotal(req.consistency?.requirement?.oneofKind || "undefined");
+
+        const result = await this.checkInternal(req, experimentsFields, forceEnablement);
+        if (result.checkedAt) {
+            await this.tokenCache.set([req.resource, result.checkedAt]);
+        }
+        return result.permitted;
+    }
+
+    private async checkInternal(
         req: v1.CheckPermissionRequest,
         experimentsFields: {
             userId: string;
@@ -119,6 +123,20 @@ export class SpiceDBAuthorizerImpl implements SpiceDBAuthorizer {
     }
 
     async writeRelationships(...updates: v1.RelationshipUpdate[]): Promise<v1.WriteRelationshipsResponse | undefined> {
+        const result = await this.writeRelationshipsInternal(...updates);
+        const writtenAt = result?.writtenAt?.token;
+        await this.tokenCache.set(
+            ...updates.map<ZedTokenCacheKV>((u) => [
+                u.relationship?.resource,
+                writtenAt, // Make sure that in case we don't get a writtenAt token here, we at least invalidate the cache
+            ]),
+        );
+        return result;
+    }
+
+    private async writeRelationshipsInternal(
+        ...updates: v1.RelationshipUpdate[]
+    ): Promise<v1.WriteRelationshipsResponse | undefined> {
         const timer = spicedbClientLatency.startTimer();
         let error: Error | undefined;
         try {
@@ -141,7 +159,19 @@ export class SpiceDBAuthorizerImpl implements SpiceDBAuthorizer {
         }
     }
 
-    async deleteRelationships(req: v1.DeleteRelationshipsRequest): Promise<DeletionResult> {
+    async deleteRelationships(req: v1.DeleteRelationshipsRequest): Promise<v1.ReadRelationshipsResponse[]> {
+        const result = await this.deleteRelationshipsInternal(req);
+        log.info(`[spicedb] Deletion result`, { result });
+        const deletedAt = result?.deletedAt;
+        if (deletedAt) {
+            await this.tokenCache.set(
+                ...result.relationships.map<ZedTokenCacheKV>((r) => [r.relationship?.resource, deletedAt]),
+            );
+        }
+        return result.relationships;
+    }
+
+    private async deleteRelationshipsInternal(req: v1.DeleteRelationshipsRequest): Promise<DeletionResult> {
         const timer = spicedbClientLatency.startTimer();
         let error: Error | undefined;
         try {
@@ -182,6 +212,7 @@ export class SpiceDBAuthorizerImpl implements SpiceDBAuthorizer {
     }
 
     async readRelationships(req: v1.ReadRelationshipsRequest): Promise<v1.ReadRelationshipsResponse[]> {
+        // pass through with given consistency/caching for now, as we only use it for reflection so far
         return tryThree("readRelationships failed.", () => this.client.readRelationships(req, this.callOptions));
     }
 
@@ -194,5 +225,116 @@ export class SpiceDBAuthorizerImpl implements SpiceDBAuthorizer {
         return (<grpc.CallOptions>{
             deadline: Date.now() + 8000,
         }) as any as grpc.Metadata;
+    }
+}
+
+type ZedTokenCacheKV = [objectRef: v1.ObjectReference | undefined, token: string | undefined];
+interface ZedTokenCache {
+    get(objectRef: v1.ObjectReference): Promise<string | undefined>;
+    set(...kvs: ZedTokenCacheKV[]): Promise<void>;
+    consistency(resourceRef: v1.ObjectReference | undefined): Promise<v1.Consistency>;
+}
+
+type ContextWithZedToken = RequestContext & { zedToken?: StoredZedToken };
+function getContext(): ContextWithZedToken {
+    return getRequestContext() as ContextWithZedToken;
+}
+
+/**
+ * This is a simple implementation of the ZedTokenCache that uses the local context to store single ZedToken per API request, which is stored in AsyncLocalStorage.
+ * To make this work we make the "assumption" that ZedTokens string (meant to be opaque) represent a timestamp which we can order. This is at least true for the MySQL datastore we are using.
+ */
+export class RequestLocalZedTokenCache implements ZedTokenCache {
+    constructor() {}
+
+    async get(objectRef: v1.ObjectReference): Promise<string | undefined> {
+        return getContext().zedToken?.token;
+    }
+
+    async set(...kvs: ZedTokenCacheKV[]) {
+        function clearZedTokenOnContext() {
+            getContext().zedToken = undefined;
+        }
+
+        const mustClearCache = kvs.some(([k, v]) => !!k && !v); // did we write a relationship without getting a writtenAt token?
+        if (mustClearCache) {
+            clearZedTokenOnContext();
+            return;
+        }
+
+        try {
+            const allTokens = [
+                ...kvs.map(([_, v]) => (!!v ? StoredZedToken.fromToken(v) : undefined)),
+                getContext().zedToken,
+            ].filter((v) => !!v) as StoredZedToken[];
+            const freshest = this.freshest(...allTokens);
+            if (freshest) {
+                getContext().zedToken = freshest;
+            }
+        } catch (err) {
+            log.warn("[spicedb] Failed to set ZedToken on context", err);
+            clearZedTokenOnContext();
+        }
+    }
+
+    async consistency(resourceRef: v1.ObjectReference | undefined): Promise<v1.Consistency> {
+        function fullyConsistent() {
+            return v1.Consistency.create({
+                requirement: {
+                    oneofKind: "fullyConsistent",
+                    fullyConsistent: true,
+                },
+            });
+        }
+        if (!resourceRef) {
+            return fullyConsistent();
+        }
+
+        const zedToken = await this.get(resourceRef);
+        if (!zedToken) {
+            return fullyConsistent();
+        }
+        return v1.Consistency.create({
+            requirement: {
+                oneofKind: "atLeastAsFresh",
+                atLeastAsFresh: v1.ZedToken.create({
+                    token: zedToken,
+                }),
+            },
+        });
+    }
+
+    protected freshest(...zedTokens: StoredZedToken[]): StoredZedToken | undefined {
+        return zedTokens.reduce<StoredZedToken | undefined>((prev, curr) => {
+            if (!prev || prev.timestamp < curr.timestamp) {
+                return curr;
+            }
+            return prev;
+        }, undefined);
+    }
+}
+
+export interface StoredZedToken {
+    token: string;
+    timestamp: number;
+}
+namespace StoredZedToken {
+    export function create(token: string, timestamp: number): StoredZedToken {
+        return { token, timestamp };
+    }
+
+    export function fromToken(token: string): StoredZedToken | undefined {
+        // following https://github.com/authzed/spicedb/blob/786555c24af98abfe3f832c94dbae5ca518dcf50/pkg/zedtoken/zedtoken.go#L64-L100
+        const decodedBytes = base64decode(token);
+        const decodedToken = DecodedZedToken.decode(Buffer.from(decodedBytes, "utf8")).v1;
+        if (!decodedToken) {
+            return undefined;
+        }
+
+        // for MySQL:
+        //  - https://github.com/authzed/spicedb/blob/main/internal/datastore/mysql/revisions.go#L182C1-L189C2
+        //  - https://github.com/authzed/spicedb/blob/786555c24af98abfe3f832c94dbae5ca518dcf50/pkg/datastore/revision/decimal.go#L53
+        const timestamp = parseInt(decodedToken.revision, 10);
+        return { token, timestamp };
     }
 }
